requirements.txt
Django>=3.0.6,<3.1.0
djangorestframework>=3.11.0,<3.12.0
celery==4.4.2
slackclient==2.5.0
djangorestframework-simplejwt==4.4.0